+++ apps/src/componentLibrary/button/Button.story.tsx
@@ -98,13 +98,6 @@ GroupOfColorsOfButtons.args = {
       onClick: () => null,
     },
     {
-      text: 'Button Secondary Purple',
-      color: buttonColors.purple,
-      type: 'secondary',
-      size: 'm',
-      onClick: () => null,
-    },
-    {
       text: 'Button Secondary Black',
       color: buttonColors.black,
       type: 'secondary',
@@ -166,14 +159,6 @@ GroupOfColorsOfButtons.args = {
       icon: {iconName: 'smile', iconStyle: 'solid'},
       color: buttonColors.white,
       type: 'primary',
-      isIconOnly: true,
-      size: 'm',
-      onClick: () => null,
-    },
-    {
-      icon: {iconName: 'smile', iconStyle: 'solid'},
-      color: buttonColors.purple,
-      type: 'secondary',
       isIconOnly: true,
       size: 'm',
       onClick: () => null,